Improvements for interpolated strings; highlighting/bm for xml comments; SCL-4385; SCL-4350; SCL-4348; SCL-4295; SCL-4223; SCL-4188; SCL-4126; some improvements for play 2 plugin; tests (cherry picked from commit c9de149)

+++ src/org/jetbrains/plugins/scala/lang/parser/parsing/xml/AttrValue.scala
@@ -4,9 +4,10 @@ package parser
 package parsing
 package xml
 
-import com.intellij.lang.PsiBuilder
 import com.intellij.psi.xml.XmlTokenType
 import builder.ScalaPsiBuilder
+import parser.util.ParserPatcher
+import com.intellij.psi.tree.TokenSet
 
 /**
 * @author Alexander Podkhalyuzin
@@ -20,22 +21,26 @@ import builder.ScalaPsiBuilder
  */
 
 object AttrValue {
+  private val VALID_ATTRIBUTE_TOKENS = TokenSet.create(XmlTokenType.XML_ATTRIBUTE_VALUE_TOKEN, XmlTokenType.XML_CHAR_ENTITY_REF)
+  
   def parse(builder: ScalaPsiBuilder): Boolean = {
     val attrValueMarker = builder.mark()
+    val patcher = ParserPatcher.getSuitablePatcher(builder)
+    
     builder.getTokenType match {
       case XmlTokenType.XML_ATTRIBUTE_VALUE_START_DELIMITER => {
         builder.advanceLexer()
-        builder.getTokenType match {
-          case XmlTokenType.XML_ATTRIBUTE_VALUE_TOKEN | XmlTokenType.XML_CHAR_ENTITY_REF => builder.advanceLexer()
-          case _ =>
+        while (VALID_ATTRIBUTE_TOKENS.contains(builder.getTokenType)) {
+          builder.advanceLexer()
         }
+        while (patcher.parse(builder)){}
         builder.getTokenType match {
           case XmlTokenType.XML_ATTRIBUTE_VALUE_END_DELIMITER => builder.advanceLexer()
           case _ => builder error ErrMsg("xml.attribute.end.expected")
         }
       }
       case _ => {
-        if (!ScalaExpr.parse(builder)) {
+        if (!ScalaExpr.parse(builder) && !patcher.parse(builder)) {
           attrValueMarker.drop()
           return false
         }
